requirements.txt
docker>=7.0.0
python-owasp-zap-v2.4>=0.0.22
pyyaml>=6.0
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0
urllib3<2.0.0
